Break accidental dependency between Pipeline and App stacks
The current behavior of creating an imported Role in the application
stack leads to creating a dependency between the Pipeline stack and the
application stack that the Pipeline deploys.

That is incorrect behavior, since the dependency implies:

> The application stack needs to be deployed before the pipeline stack
> can be deployed.

But in actuality, the pipeline stack will create the application stack,
so there is no such ordering dependency.

1. I think the behavior that is causing the stack dependency caused by
using an imported role from another stack to appear is a bug, and should
be fixed.
2. For now, I've worked around it by creating the imported Role not in
the scope of the application stack, but in the scope of the pipeline
stack.