Add Docker assets
Add an ECS Container running a Docker Asset container.

As predicted, this unearthed some changes to the bootstrapping
stack (missing permissions) and some problems in the `cdk-assets`
tool (which had already been fixed in the CLI PR, but not in
the `convmode` PR).

This change introduces a local development copy of the application
(`Dev-Meerkats-DDBStack` et. al.) and a special script to deploy
them (`deploy-local.sh`) since the CLI doesn't run the `cdk-assets`
tool yet, which is necessary for local deployments to work.

The API Gateway has been removed and replaced with an Application Load
Balancer, since API Gateway cannot forward traffic to ECS services
easily, or must do so always by means of an ALB or NLB. Since ALB
can also hit Lambdas, might as well replace APIGW with an ALB.