chore: rewrite to use new Stage construct of 1.45.0
`Stage` replaces the previous use of `App` to group application stacks
together in a deployable unit. `Stage` was introduced in CDK `1.45.0`,
and this module now requires that version of CDK at minimum.

In 1.45.0, we split the asset publishing role into 2 roles, so
an update of the bootstrap stack is required.

Add a slew of unit tests to enshrine the current behavior, flushing
out a slew of minor misbehaviors and missing features:

- Stack templates are no longer published as assets, because the
  CodePipeline CloudFormation action will read the template from
  the artifact bucket.

- More control over run orders, add option for manual approval
  in between.

- Multiple independent stacks are deployed in parallel.

- Don't unnecessarily assume privileged mode on file asset publishing
  CodeBuild jobs.

- Allow control over CDK CLI version used in pipeline.

- NPM and Yarn builds support running a project that's in a subdirectory
  of the git repo.

- Validation shellscripts now run with `set -eu`, make that
  configurable.

- CodeBuild projects have deterministic names if the pipeline is
  given a deterministic name, for easier recognizability.